backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "google-adk>=0.5.0",
    "sqlalchemy>=2.0.41",
    "supabase>=2.15.1",
]
